refactor: 「ユーザの手によってのアレンジ」を実際に反映する手続きを、クラスとして分割。
