+++ src/domain/diagram/import/ImportDiagramCandidate.ts
@@ -1,9 +1,6 @@
-import UserArrangeOfImportDiagram from "@/domain/diagram/import/userarrange/UserArrangeOfImportDiagram";
-import ConflictNameBehavior from "@/domain/diagram/import/conflictname/ConflictNameBehavior";
 import Resource from "@/domain/resource/Resource";
 import Diagram from "@/domain/diagram/Diagram";
 import Product from "@/domain/product/Product";
-import { BehaviorWhenNameConflict } from "@/domain/diagram/import/userarrange/BehaviorWhenNameConflict";
 
 export default class ImportDiagramCandidate {
     constructor(
@@ -26,53 +23,6 @@ export default class ImportDiagramCandidate {
         const newResources = this.useResources
             .map(r => r.renewId(-r.resourceId))
         return new ImportDiagramCandidate(newDiaglam, newResources);
-    }
-
-    public arrangeOf(
-        userArrange: UserArrangeOfImportDiagram,
-        product: Product
-    ): ImportDiagramCandidate | null {
-        let modifiedDiagram = this.diagram;
-        let modifiedResources = this.useResources;
-
-        for (const colidedResourceName of userArrange.resourceNamesOfColided) {
-            const targetResouce = modifiedResources
-                .find(r => r.resourceId === colidedResourceName.sourceId) as Resource;
-            const sameResource = product.resources.getSameOf(targetResouce) as Resource;
-
-            const behavior = colidedResourceName.behavior;
-            if (behavior === BehaviorWhenNameConflict.既存) {
-                modifiedDiagram = modifiedDiagram.replaceOf(targetResouce, sameResource);
-                modifiedResources = modifiedResources
-                    .filter(r => r.resourceId !== targetResouce.resourceId);
-            }
-            if (behavior === BehaviorWhenNameConflict.置換) {
-                const replacedIdResource = targetResouce.renewId(sameResource.resourceId);
-                modifiedResources = modifiedResources
-                    .filter(r => r.resourceId !== targetResouce.resourceId);
-                modifiedResources.push(replacedIdResource);
-                modifiedDiagram = modifiedDiagram
-                    .replaceOf(targetResouce, replacedIdResource);
-            }
-            if (behavior === BehaviorWhenNameConflict.別名) {
-                const renamedResource = targetResouce
-                    .withName(colidedResourceName.destinationName);
-                modifiedResources = modifiedResources
-                    .filter(r => r.resourceId !== targetResouce.resourceId);
-                modifiedResources.push(renamedResource);
-            }
-        }
-
-        if (userArrange.isColidedDiagramName()) {
-            const colidedDiagramName = userArrange.diagramNamesOfColided as ConflictNameBehavior;
-            if (colidedDiagramName.behavior === BehaviorWhenNameConflict.既存) return null; // 入力からは入ってこない前提。「既存」というなら「Importしない」と同義。
-            if (colidedDiagramName.behavior === BehaviorWhenNameConflict.別名) {
-                modifiedDiagram = modifiedDiagram
-                    .renameOf(colidedDiagramName.destinationName);
-            }
-        }
-
-        return new ImportDiagramCandidate(modifiedDiagram, modifiedResources);
     }
 
     public mergeOf(product: Product): Product {
